// src/types.d.ts
export declare type InstanceIOValue = Vec1 | Vec2 | Vec3 | Vec4;
export declare type UniformIOValue = number | InstanceIOValue | Mat3x3 | Mat4x4 | Float32Array | Texture | number[];
export declare enum InstanceBlockIndex {
    ONE = 1,
    TWO = 2,
    THREE = 3,
    FOUR = 4
}
export declare enum InstanceAttributeSize {
    ONE = 1,
    TWO = 2,
    THREE = 3,
    FOUR = 4,
    ATLAS = 99
}
export declare const instanceAttributeSizeFloatCount: {
    [key: number]: number;
};
export declare enum UniformSize {
    ONE = 1,
    TWO = 2,
    THREE = 3,
    FOUR = 4,
    MATRIX3 = 9,
    MATRIX4 = 16,
    VEC4_ARRAY = 98,
    ATLAS = 99
}
export declare enum VertexAttributeSize {
    ONE = 1,
    TWO = 2,
    THREE = 3,
    FOUR = 4
}
export declare enum TextureSize {
    _2 = 2,
    _4 = 4,
    _8 = 8,
    _16 = 16,
    _32 = 32,
    _64 = 64,
    _128 = 128,
    _256 = 256,
    _512 = 512,
    _1024 = 1024,
    _2048 = 2048,
    _4096 = 4096
}
export declare enum ResourceType {
    ATLAS = 0,
    FONT = 1
}
// ...
export interface IInstanceAttribute<T extends Instance> {
    resource?: {
        key(): string;
        name: string;
        shaderInjection?: ShaderInjectionTarget;
    };
    block?: number;
    blockIndex?: InstanceBlockIndex;
    childAttributes?: IInstanceAttribute<T>[];
    easing?: IAutoEasingMethod<Vec>;
    name: string;
    parentAttribute?: IInstanceAttribute<T>;
    qualifier?: string;
    size?: InstanceAttributeSize;
    update(instance: T): InstanceIOValue;
}
export interface IInstanceAttributeInternal<T extends Instance> extends IInstanceAttribute<T> {
    uid: number;
    packUID?: number;
    bufferAttribute: Attribute;
}
export interface IResourceInstanceAttribute<T extends Instance> extends IInstanceAttribute<T> {
    resource: {
        key(): string;
        name: string;
        shaderInjection?: ShaderInjectionTarget;
    };
}
// ...
export interface IValueInstanceAttribute<T extends Instance> extends IInstanceAttribute<T> {
    atlas: undefined;
}
export declare enum ShaderInjectionTarget {
    VERTEX = 1,
    FRAGMENT = 2,
    ALL = 3
}
export interface IUniform {
    shaderInjection?: ShaderInjectionTarget;
    name: string;
    size: UniformSize;
    qualifier?: string;
    update(uniform: IUniform): UniformIOValue;
}